Refactor API a bit
Split up the structs for the live data into multiple files to make it a little more manageable
