+++ Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 members = [".", "api"]
 
 [dependencies]
-mlb-api = {path = "api/", version = "0.0.3"}
+mlb-api = {path = "api/", version = "0.0.4"}
 chrono = "0.4"
 tui = "0.14"
 termion = "1.5"
