+++ api/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
 name = "mlb-api"
-version = "0.0.3"
+version = "0.0.4"
 authors = ["Andrew Schneider <andjschneider@gmail.com>"]
 edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 chrono = "0.4"
